+++ bin/node/src/service.rs
@@ -7,10 +7,10 @@ use std::{
 
 use fake_runtime_api::fake_runtime::RuntimeApi;
 use finality_aleph::{
-    build_network, get_aleph_block_import, run_validator_node, AlephConfig, AllBlockMetrics,
-    BlockImporter, BuildNetworkOutput, ChannelProvider, FavouriteSelectChainProvider,
-    Justification, JustificationTranslator, MillisecsPerBlock, RateLimiterConfig,
-    RedirectingBlockImport, SessionPeriod, SubstrateChainStatus, SyncOracle, ValidatorAddressCache,
+    build_network, get_aleph_block_import, run_validator_node, AlephConfig, BlockImporter,
+    BuildNetworkOutput, ChannelProvider, FavouriteSelectChainProvider, Justification,
+    JustificationTranslator, MillisecsPerBlock, RateLimiterConfig, RedirectingBlockImport,
+    SessionPeriod, SubstrateChainStatus, SyncOracle, ValidatorAddressCache,
 };
 use log::warn;
 use pallet_aleph_runtime_api::AlephSessionApi;
